move network resources
