+++ lib/stack/azurerm-playground-stack.ts
@@ -1,17 +1,11 @@
 import { Construct } from 'constructs';
 import { TerraformStack, TerraformOutput } from 'cdktf';
 import { provider } from '@cdktf/provider-azurerm';
-import { UserAssignedIdentity } from '../construct/azurerm/user-assigned-identity';
+// import { UserAssignedIdentity } from '../construct/azurerm/user-assigned-identity';
 // import { RoleAssignment } from '../construct/azurerm/role-assignment';
-import { KeyVault } from '../construct/azurerm/key-vault';
+// import { KeyVault } from '../construct/azurerm/key-vault';
 import { ResourceGroup } from '../construct/azurerm/resource-group';
 import { Iothub } from '../construct/azurerm/iothub';
-import { VirtualNetwork } from '../construct/azurerm/virtual-network';
-import { Subnet } from '../construct/azurerm/subnet';
-import { VirtualMachine } from '../construct/azurerm/virtual-machine';
-import { BastionHost } from '../construct/azurerm/bastion';
-import { PrivateDnsZone } from '../construct/azurerm/private-dns-zone';
-import { PrivateEndpoint } from '../construct/azurerm/private-endpoint';
 // import { MonitorDiagnosticSetting } from '../construct/azurerm/monitor-diagnostic-setting';
 import { MonitorWorkspace } from '../construct/azurerm/monitor-workspace';
 import { EventgridNamespace } from '../construct/azurerm/eventgrid-namespace';
@@ -43,19 +37,6 @@ export interface AzurermPlaygroundStackProps {
     skuName: string;
     capacity: number;
   };
-  virtualNetwork?: {};
-  subnet?: {
-    name: string;
-    addressPrefixes: string[];
-  }[];
-  virtualMachine?: {
-    vmSize: string;
-  };
-  bastionHost?: {
-    sku: string;
-  };
-  privateDnsZone?: {};
-  privateEndpoint?: {};
   monitorDiagnosticSetting?: {};
   monitorWorkspace?: {};
   eventgridNamespace?: {};
@@ -135,29 +116,6 @@ export const prodAzurermPlaygroundStackProps: AzurermPlaygroundStackProps = {
   keyVault: {
     skuName: 'standard',
   },
-  virtualNetwork: {},
-  subnet: [
-    {
-      name: 'VmSubnet',
-      addressPrefixes: ['10.240.0.0/16'],
-    },
-    {
-      name: 'AzureBastionSubnet',
-      addressPrefixes: ['10.241.0.0/16'],
-    },
-    {
-      name: 'PrivateEndpointSubnet',
-      addressPrefixes: ['10.242.0.0/16'],
-    },
-  ],
-  virtualMachine: {
-    vmSize: 'Standard_DS2_v2',
-  },
-  bastionHost: {
-    sku: 'Basic',
-  },
-  privateDnsZone: {},
-  privateEndpoint: {},
 };
 
 export class AzurermPlaygroundStack extends TerraformStack {
@@ -194,31 +152,31 @@ export class AzurermPlaygroundStack extends TerraformStack {
       value: resourceGroup.resourceGroup.name,
     });
 
-    let userAssignedIdentity: UserAssignedIdentity | undefined = undefined;
-    if (props.userAssignedIdentity) {
-      userAssignedIdentity = new UserAssignedIdentity(
-        this,
-        `UserAssignedIdentity`,
-        {
-          name: `uai-${props.name}`,
-          location: props.location,
-          tags: props.tags,
-          resourceGroupName: resourceGroup.resourceGroup.name,
-        },
-      );
-    }
+    // let userAssignedIdentity: UserAssignedIdentity | undefined = undefined;
+    // if (props.userAssignedIdentity) {
+    //   userAssignedIdentity = new UserAssignedIdentity(
+    //     this,
+    //     `UserAssignedIdentity`,
+    //     {
+    //       name: `uai-${props.name}`,
+    //       location: props.location,
+    //       tags: props.tags,
+    //       resourceGroupName: resourceGroup.resourceGroup.name,
+    //     },
+    //   );
+    // }
 
-    let keyVault: KeyVault | undefined = undefined;
-    if (props.keyVault) {
-      keyVault = new KeyVault(this, `KeyVault`, {
-        name: convertName(`kv-${props.name}`, 24),
-        location: props.location,
-        tags: props.tags,
-        resourceGroupName: resourceGroup.resourceGroup.name,
-        skuName: props.keyVault.skuName,
-        purgeProtectionEnabled: false,
-      });
-    }
+    // let keyVault: KeyVault | undefined = undefined;
+    // if (props.keyVault) {
+    //   keyVault = new KeyVault(this, `KeyVault`, {
+    //     name: convertName(`kv-${props.name}`, 24),
+    //     location: props.location,
+    //     tags: props.tags,
+    //     resourceGroupName: resourceGroup.resourceGroup.name,
+    //     skuName: props.keyVault.skuName,
+    //     purgeProtectionEnabled: false,
+    //   });
+    // }
 
     if (props.iothub) {
       new Iothub(this, `Iothub`, {
@@ -228,144 +186,6 @@ export class AzurermPlaygroundStack extends TerraformStack {
         resourceGroupName: resourceGroup.resourceGroup.name,
         skuName: props.iothub.skuName,
         capacity: props.iothub.capacity,
-      });
-    }
-
-    let virtualNetwork: VirtualNetwork | undefined = undefined;
-    if (props.virtualNetwork) {
-      virtualNetwork = new VirtualNetwork(this, `VirtualNetwork`, {
-        name: `vnet-${props.name}`,
-        location: props.location,
-        tags: props.tags,
-        resourceGroupName: resourceGroup.resourceGroup.name,
-        addressSpace: ['10.0.0.0/8'],
-      });
-    }
-
-    let subnet: Subnet | undefined = undefined;
-    if (props.subnet && virtualNetwork) {
-      subnet = new Subnet(this, `Subnet`, {
-        subnets: props.subnet.map((subnet) => {
-          return {
-            name: subnet.name,
-            resourceGroupName: resourceGroup.resourceGroup.name,
-            virtualNetworkName: virtualNetwork.virtualNetwork.name,
-            addressPrefixes: subnet.addressPrefixes,
-          };
-        }),
-      });
-    }
-
-    if (props.virtualMachine && subnet) {
-      new VirtualMachine(this, `VirtualMachine`, {
-        name: `vm-${props.name}`,
-        location: props.location,
-        tags: props.tags,
-        resourceGroupName: resourceGroup.resourceGroup.name,
-        subnetId: subnet.subnets[0].id,
-        vmSize: props.virtualMachine.vmSize,
-        identityIds: userAssignedIdentity
-          ? [userAssignedIdentity.userAssignedIdentity.id]
-          : undefined,
-      });
-    }
-
-    if (props.bastionHost && subnet) {
-      new BastionHost(this, `BastionHost`, {
-        name: `bastion-${props.name}`,
-        location: props.location,
-        tags: props.tags,
-        resourceGroupName: resourceGroup.resourceGroup.name,
-        sku: props.bastionHost.sku,
-        subnetId: subnet.subnets[1].id,
-      });
-    }
-
-    // Create Private DNS Zones for different services
-    const privateDnsZones: { [key: string]: PrivateDnsZone } = {};
-    if (props.privateDnsZone && virtualNetwork) {
-      const dnsZoneConfigs = [
-        { id: 'OpenAi', name: 'privatelink.openai.azure.com' },
-        { id: 'StorageAccount', name: 'privatelink.blob.core.windows.net' },
-        { id: 'KeyVault', name: 'privatelink.vaultcore.azure.net' },
-        { id: 'ContainerRegistry', name: 'privatelink.azurecr.io' },
-      ];
-
-      dnsZoneConfigs.forEach((config) => {
-        privateDnsZones[config.id] = new PrivateDnsZone(
-          this,
-          `PrivateDnsZone${config.id}`,
-          {
-            name: config.name,
-            tags: props.tags,
-            resourceGroupName: resourceGroup.resourceGroup.name,
-            virtualNetworkId: virtualNetwork.virtualNetwork.id,
-          },
-        );
-      });
-    }
-
-    // Create Private Endpoints
-    if (props.privateEndpoint && subnet) {
-      // Create Private Endpoints for AI Services
-      if (privateDnsZones['OpenAi']) {
-        // aiServicesArray.forEach((aiService) => {
-        //   new PrivateEndpoint(
-        //     this,
-        //     `PrivateEndpoint-${aiService.aiServices.name}`,
-        //     {
-        //       name: `pe-${aiService.aiServices.name}`,
-        //       location: props.location,
-        //       tags: props.tags,
-        //       resourceGroupName: resourceGroup.resourceGroup.name,
-        //       subnetId: subnet.subnets[2].id,
-        //       privateConnectionResourceId: aiService.aiServices.id,
-        //       subresourceNames: ['account'],
-        //       privateDnsZoneIds: [privateDnsZones['OpenAi'].privateDnsZone.id],
-        //     },
-        //   );
-        // });
-      }
-
-      // Create Private Endpoints for other services
-      const endpointConfigs = [
-        // {
-        //   id: 'StorageAccount',
-        //   condition: !!storageAccount,
-        //   resourceId: storageAccount?.storageAccount.id,
-        //   subresource: 'blob',
-        // },
-        {
-          id: 'KeyVault',
-          condition: !!keyVault,
-          resourceId: keyVault?.keyVault.id,
-          subresource: 'vault',
-        },
-        // {
-        //   id: 'ContainerRegistry',
-        //   condition: !!containerRegistry,
-        //   resourceId: containerRegistry?.containerRegistry.id,
-        //   subresource: 'registry',
-        // },
-      ];
-
-      endpointConfigs.forEach((config) => {
-        if (
-          config.condition &&
-          privateDnsZones[config.id] &&
-          config.resourceId
-        ) {
-          new PrivateEndpoint(this, `PrivateEndpoint${config.id}`, {
-            name: `pe-${config.id.toLowerCase()}-${props.name}`,
-            location: props.location,
-            tags: props.tags,
-            resourceGroupName: resourceGroup.resourceGroup.name,
-            subnetId: subnet.subnets[2].id,
-            privateConnectionResourceId: config.resourceId,
-            subresourceNames: [config.subresource],
-            privateDnsZoneIds: [privateDnsZones[config.id].privateDnsZone.id],
-          });
-        }
       });
     }
